requirements.txt
odfpy>=1.4.1
python-docx>=0.8.11
mammoth>=1.6.0
docx2txt>=0.8
PyYAML>=6.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0 
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0 